smp: done!

+++ os/lock.h
@@ -5,7 +5,7 @@
 
 // Mutual exclusion lock.
 struct spinlock {
-	uint locked; // Is the lock held?
+	uint64 locked; // Is the lock held?
 
 	// For debugging:
 	char *name; // Name of lock.
